Updated protected routes logic and added error/loading pages
