+++ frontend/src/common/authentication/LoginRedirect.tsx
@@ -6,6 +6,7 @@ import * as routes from "../../routes/constants";
 import { HomePage } from "../../features/homePage/HomePage";
 import { getUserContext } from "../../features/manageProfile/apiManager/manageProfileAPI";
 import { UserContextType } from "./types";
+import { Loading } from "../pages/Loading";
 
 /*
  * Redirects user to their correct page after loading their
@@ -89,11 +90,7 @@ export const LoginRedirect = () => {
   }, [isLoading]);
 
   if (isLoading) {
-    return (
-      <>
-        <span>Loading...</span>
-      </>
-    );
+    return <Loading />;
   }
 
   return (
