+++ frontend/src/common/pages/UnexpectedError.tsx
@@ -1,23 +1,26 @@
-import { Button, Grid, Typography } from "@mui/material";
+import { Box, Button, Grid, Typography } from "@mui/material";
 import { useNavigate } from "react-router-dom";
 
-export const UnexpectedError = () => {
+export const UnexpectedError = ({ error }: { error: string }) => {
   const navigate = useNavigate();
   return (
-    <Grid
-      container
-      direction="row"
-      justifyContent="center"
-      alignItems="center"
-      height="calc(100vh - 100px)"
-    >
-      <Grid container item xs={6} justifyContent="center">
-        <Typography variant="h4" align="center">
-          Unexpected Error
-        </Typography>
-        <Typography variant="h4" align="center" margin={"20px"}>
-          An unexpected error occured. Click the button below to go back
-        </Typography>
+    <Box height="calc(100vh - 100px)" paddingTop="24px">
+      <Grid
+        container
+        direction="row"
+        justifyContent="center"
+        alignItems="center"
+      >
+        <Grid container item xs={12} justifyContent="center">
+          <Typography variant="h4" align="center">
+            Error:
+          </Typography>
+        </Grid>
+        <Grid container item xs={12} justifyContent="center">
+          <Typography variant="h4" align="center" margin={"20px"}>
+            {error}
+          </Typography>
+        </Grid>
 
         <Grid item>
           <Button
@@ -30,6 +33,6 @@ export const UnexpectedError = () => {
           </Button>
         </Grid>
       </Grid>
-    </Grid>
+    </Box>
   );
 };
