+++ frontend/src/features/manageProfile/apiManager/hooks.ts
@@ -1,10 +1,25 @@
 import { useQuery } from "@tanstack/react-query";
-import { getCompanyInfo } from "./manageProfileAPI";
+import { getCompanyInfo, getUserContext } from "./manageProfileAPI";
 
 export const useCompanyInfoQuery = () => {
   return useQuery({
     queryKey: ["companyInfo"],
     queryFn: getCompanyInfo,
+    retry: false,
+  });
+};
+
+export const useUserContext = () => {
+  return useQuery({
+    queryKey: ["userContext"],
+    queryFn: getUserContext,
+    onSuccess: (userContextResponseBody) => {
+      console.log(userContextResponseBody);
+      sessionStorage.setItem(
+        "onroutebc.user.context",
+        JSON.stringify(userContextResponseBody)
+      );
+    },
     retry: false,
   });
 };
