+++ frontend/src/features/manageProfile/components/dashboard/ManageProfilesDashboard.tsx
@@ -6,6 +6,7 @@ import { AxiosError } from "axios";
 import { Unauthorized } from "../../../../common/pages/Unauthorized";
 import { getCompanyInfo } from "../../apiManager/manageProfileAPI";
 import { Loading } from "../../../../common/pages/Loading";
+import { UnexpectedError } from "../../../../common/pages/UnexpectedError";
 
 export const ManageProfilesDashboard = React.memo(() => {
   const {
@@ -30,7 +31,7 @@ export const ManageProfilesDashboard = React.memo(() => {
       if (error.response?.status === 401) {
         return <Unauthorized />;
       }
-      return <span>Error: {error.message}</span>;
+      return <UnexpectedError error={error.message} />;
     }
   }
